API Reference

H2Trees.AggregateMode

H2Trees.AggregateTranslateMode

H2Trees.ParentUpwardsIterator

H2Trees.LevelIterator

H2Trees.SameLevelIterator

H2Trees.boundingbox

H2Trees.cornerpoints

H2Trees.findleafnode

H2Trees.leveltolevelid

H2Trees.minimumlevel

H2Trees.parentcenterminuschildcenter

H2Trees.values

H2Trees.AggregateMode — Type

AggregateMode <: FarMulMode

This mode uses AggregatePlan and DisaggregateTranslatePlan to perform the farmultiplication.

H2Trees.AggregateTranslateMode — Type

AggregateTranslateMode <: FarMulMode

This mode uses AggregateTranslatePlan and DisaggregatePlan to perform the farmultiplication.

H2Trees.ParentUpwardsIterator — Type

struct ParentUpwardsIterator{T}

ParentUpwardsIterator is an iterator that iterates over all parent nodes of a given node in a tree until the root is reached. The last node is the node 0.

Example

for node in DepthFirstIterator(tree, root(tree))
    println("Node: ", node)
    for parent in ParentUpwardsIterator(tree, node)
        println("	 Parent: ", parent)
    end
end

Fields

tree::T: The tree.

node::Int: The node over which parents is iterated.

H2Trees.LevelIterator — Method

LevelIterator(tree, level::Int)

Return an iterator over the nodes at the specified level in the tree.

Arguments

tree: The tree object.

level: The level at which to iterate.

Returns

An iterator over the nodes at the specified level.

H2Trees.SameLevelIterator — Method

SameLevelIterator(tree, node::Int)

Returns an iterator over the nodes at the same level as node in the tree.

Arguments

tree: The tree structure.

node: The node for which to find the same level nodes.

Returns

An iterator over the nodes at the same level as node.

H2Trees.boundingbox — Method

boundingbox(points::Vector{SVector{D, T}})

Returns halfsize and center of bounding box of points. The halfsize is the half of the length of the edge of the bounding box.

H2Trees.cornerpoints — Method

cornerpoints(tree::TwoNTree{N,D,T}, node::Int, i)

Return the corner point of a given node in an N-dimensional TwoNTree.

Arguments

tree::TwoNTree{N,D,T}: The tree.

node::Int: The index of the node.

i: The corner point index (1 til 2^N).

Returns

The corner point coordinates as a SVector.

H2Trees.findleafnode — Method

findleafnode(tree, value::Int)

Find the leaf node in the given tree that contains the specified value.

Arguments

tree: The tree to search in.

value: The value to search for.

Returns

The leaf node that contains the value, or 0 if not found.

H2Trees.leveltolevelid — Method

leveltolevelid(tree, level::Int)

Converts a level in the tree to its corresponding level ID. This is relevant since the first level might not be level 1.

Arguments

tree: The tree object.

level: The level to convert.

Returns

The level ID corresponding to the given level.

H2Trees.minimumlevel — Method

minimumlevel(tree)

Get the minimum level of a tree, which is the level of the root node. This is not necessarily the level 1.

Arguments

tree: The tree.

Returns

The minimum level of the tree.

H2Trees.parentcenterminuschildcenter — Method

parentcenterminuschildcenter(tree::TwoNTree{N,D,T}, child::Int) where {D,T}

Calculate the difference r_p-r_c between the center of the parent r_p and the center of the child node r_c.

Arguments

tree::TwoNTree{N,D,T}

child::Int: The index of the child node.

Returns

SVector{N,T}: The difference between the center of the parent node and the center of the child node.

H2Trees.values — Method

values(tree, node::Int)

Returns the values stored in the given node of the tree. If the node is a leaf node, it returns the values directly. Otherwise, it recursively collects the values from all the leaf nodes in the subtree rooted at the given node.

Arguments

tree: The H2 tree.

node::Int: The index of the node.

Returns

An array of values stored in the given node or its subtree.